Music player with basic functionalities.
Here is a music player code that is implemented using Python Tkinter. It has basic functionality i.e. play, pause, next, previous, search, and find a song from a directory. In this version, the user has to modify the directory address in the code and cannot do the same from the GUI window. Another lack of this music player is a scale widget is not automatically updating and only use to skip the song.