"YOUR TODAY"
Now a days online schooling and colleges are going on. And our major task in this is find out the the next lecture and meeting link for it.
So during the free time I tried to make something that reduce this 'HARD' task for us. (generally students) 

Your today is the python script that will automate this task for you.
It is very basic version and not implemented GUI. The other thing is that it is not that much robust so you have to input your own time table inside a code. 
In short If you want to change time table then you have to change it INSIDE A CODE.  :(             (SORRY FOR THIS ) 

working:  After modifying the time table and attached link according to the time table, you just need to run this python program inside any editor or you can use 'PYTHON IDLE'.
-> Next step is ....
       DO NOTHING.
It will automatically open particular link in your default browser.